fix: address PR review feedback on pluggable kernel architecture
- OneEuro.cs: Remove unnecessary (float) cast in ComputeAlpha (precision loss)
- OneEuro3D.cs: Add namespace ESPresense.Filtering and mark fields readonly
- IterativeNadarayaWatsonMultilateralizer.cs: Refactor complex confidence calculation into readable intermediate variables with comments
- GaussianKernel.cs: Add negative distance validation (consistent with other kernels)
- State.cs: Fix null reference warnings (CS8602) using local variables and null-forgiving operator
- ConfigTests.cs: Update tests to reflect new NadarayaWatsonConfig structure (Kernel property instead of Bandwidth)

These changes improve code quality, maintainability, and address compiler warnings while preserving the pluggable kernel architecture.

+++ src/Filtering/OneEuro.cs
@@ -81,7 +81,7 @@ public class OneEuroFilter
     private static double ComputeAlpha(double cutoffFreq, double dt)
     {
         double tau = 1.0 / (2.0 * Math.PI * cutoffFreq);
-        return (float)(1.0 / (1.0 + tau / dt));
+        return 1.0 / (1.0 + tau / dt);
     }
 }
